# feat: optimize cache pointer performance with preemptive multi-pointer caching